+++ log.txt
@@ -1,1 +1,3 @@
-2022-09-14 01:43:04 : 12 * 12 = 144
+2022-09-14 02:36:29 : 12 * 12 = 144
+2022-09-14 02:36:36 : 144 / 213 = 0.676056
+2022-09-14 02:36:47 : 2 ^ 255555 = 1.#INF
